Rename DataEntryViewArguments to FormSectionViewModel. DataEntryViewArguments will be created specifically for DataEntryView and not shared between Form and DataEntry screens
